fixed issue in new distributed implementation from review and first testing, added new lock manager

+++ distributed/src/main/java/com/orientechnologies/orient/server/distributed/impl/ODatabaseDocumentDistributed.java
@@ -637,6 +637,7 @@ public class ODatabaseDocumentDistributed extends ODatabaseDocumentEmbedded {
   public void rollback2pc(ODistributedRequestId transactionId) {
     ODistributedDatabase localDistributedDatabase = getStorageDistributed().getLocalDistributedDatabase();
     ODistributedTxContext txContext = localDistributedDatabase.popTxContext(transactionId);
-    txContext.destroy();
+    if (txContext != null)
+      txContext.destroy();
   }
 }
